add unique index to table. update DAO method Update money amount to BigDecimal

+++ src/main/java/com/taskforce/moneyapp/dao/UserDAO.java
@@ -9,7 +9,7 @@ public interface UserDAO {
     List<User> getAllUsers();
     User getUserById(long userId);
     User getUserByName(String userName);
-    int insertUser(User user);
+    long insertUser(User user);
     int updateUser(User user);
     int deleteUser(long userId);
